pyproject.toml
[project]
name = "mini-ml"
version = "0.1.0"
description = "A small project for model rewriting"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "onnxruntime-gpu>=1.22.0",
    "opencv-python>=4.12.0.88",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "rich>=14.1.0",
    "safetensors>=0.6.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.8",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu129", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu129", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true

[project.scripts]
train="mini_ml.core.train:main"